feat: add Forum Explorer
Adds the new forum explorer and registers it as the default page navigated to on /

The forum explorer serves as the hub where users will find forums they are looking for or find new forums that might interest them. From the explorer they can navigate to the forum's page. This commit adds a bare bones implementation that just lists all forums.