refactor: Improve existing forms
Forms previously checked validity onSubmit, now the submit button remains disabled until the form is valid allowing us to remove this check.

The component previously make two separate calls to reset the form and reset the 'submitted' flag so that validation errors would not trigger immediately on the new empty form, this was move to a single call to NgForm.resetForm() in the template.