refactor: add router resolvers for navigation to forums
Adds two router resolvers whose purpose is to heat up the cache in the ForumService and PostService. This means that subsequent requests to the services initially response synchronously from the cache allowing the pre-loading effect of router resolvers without changing previously written code in the components that now benefit from it.